2nd commit

+++ api/utils/imageminMiddleware.js
@@ -6,17 +6,21 @@ const { promises: fsPromises } = fs;
 const CustomError = require("./errors");
 
 async function minifyImage(req, res, next) {
-  try {
-    await imagemin([req.file.path], {
-      destination: "public/images",
-      plugins: [imageminJpegTran()],
-    });
-    await fsPromises.unlink(req.file.path);
-    req.file.destination = "images";
-    req.file.path = `images/${req.file.filename}`;
+  if (!req.file) {
     next();
-  } catch (error) {
-    next(new CustomError(500, "Image minification error"));
+  } else {
+    try {
+      await imagemin([req.file.path], {
+        destination: "public/images",
+        plugins: [imageminJpegTran()],
+      });
+      await fsPromises.unlink(req.file.path);
+      req.file.destination = "images";
+      req.file.path = `images/${req.file.filename}`;
+      next();
+    } catch (error) {
+      next(new CustomError(500, "Image minification error"));
+    }
   }
 }
